Fix leaking resource
In my tests I'm getting "A fetch response body (rid 13) was created during the test, but not consumed during the test. Consume or close the response body `ReadableStream`, e.g `await resp.text()` or `await resp.body.cancel()"

Tracked it down to calling upload on storage and noticed that in the "ok" case, the body is never consumed there.